refactor: move download single file functions to core
